int overflow

+++ src/main/java/ch/hslu/oop/sw06/Calculator.java
@@ -3,8 +3,8 @@ package ch.hslu.oop.sw06;
 public class Calculator implements Icalculator {
 
     @Override
-    public int addition(final int x, final int y) {
+    public long addition(final int x, final int y) {
         return x + y;
     }
-    
+   
 }
